+++ Changelog.md
@@ -4,6 +4,7 @@
 - Fix Path Traversal on windows
 - Use AsRef<str> more instead of Display more
 - Add a serve path to Serve Static
+- Serve index from serve path
 
 # 1.1.0